Semana_03_Pipeline_Grafico/semana_3_2_zbuffer_depth_testing/semana_3_2_zbuffer_depth_testing/python/requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=8.0.0
